+++ resources/views/bootstrap-5/includes/filter-pills.blade.php
@@ -9,7 +9,7 @@
                         wire:key="filter-pill-{{ $key }}"
                         class="badge rounded-pill bg-info d-inline-flex align-items-center"
                     >
-                        {{ $filterNames[$key] ?? collect($this->columns())->pluck('text', 'column')->get($key, ucwords(strtr($key, ['_' => ' ', '-' => ' ']))) }}:
+                        {{ $filterNames[$key] ?? collect($this->columns())->pluck('text', 'column')->get($key, isset($customFilters[$key]) && property_exists($customFilters[$key], 'name') ? $customFilters[$key]->name : ucwords(strtr($key, ['_' => ' ', '-' => ' ']))) }}:
                         @if(isset($customFilters[$key]) && method_exists($customFilters[$key], 'options'))
                             @if(is_array($value))
                                 @foreach($value as $selectedValue)