+++ resources/views/bootstrap-5/includes/filter-type-multiselect.blade.php
@@ -1,5 +1,15 @@
+<div class="form-check p-0">
+    <input
+        onclick="event.stopPropagation();"
+        type="checkbox"
+        id="filter-{{$key}}-select-all"
+        wire:input="selectAllFilters('{{ $key }}')"
+        {{ count($filters[$key]) === count($filter->options()) ? 'checked' : ''}}
+    >
+    <label class="form-check-label" for="filter-{{$key}}-select-all">@lang('All')</label>
+</div>
 @foreach($filter->options() as $optionKey => $value)
-    <div class="form-check" wire:key="filter-{{ $key }}-multiselect-{{ $optionKey }}">
+    <div class="form-check p-0" wire:key="filter-{{ $key }}-multiselect-{{ $optionKey }}">
         <input
             onclick="event.stopPropagation();"
             type="checkbox"