+++ resources/views/tailwind/includes/filter-type-multiselect.blade.php
@@ -1,4 +1,13 @@
 <div class="mt-1 relative rounded-md">
+    <div>
+        <input
+            type="checkbox"
+            id="filter-{{$key}}-select-all"
+            wire:input="selectAllFilters('{{ $key }}')"
+            {{ count($filters[$key]) === count($filter->options()) ? 'checked' : ''}}
+        >
+        <label for="filter-{{$key}}-select-all">@lang('All')</label>
+    </div>
     @foreach($filter->options() as $optionKey => $value)
         <div wire:key="filter-{{ $key }}-multiselect-{{ $optionKey }}">
             <input